+++ app/Http/Controllers/PageController.php
@@ -25,6 +25,8 @@ class PageController extends Controller
      */
     public function index(Request $request)
     {
+        // Only Admins
+        if(!$this->hasrole('Admin')) { return redirect('/'); }
         if(isset($request->q)) {
             $pages = Page::where('name', 'LIKE', '%'.$request->q.'%')->orWhere('content', 'LIKE', '%'.$request->q.'%')->orderBy('updated_at', 'desc')->paginate(100);
         } else {
@@ -35,6 +37,9 @@ class PageController extends Controller
 
     public function store(Request $request)
     {
+        // Only Admins
+        if(!$this->hasrole('Admin')) { return redirect('/'); }
+
         $this->validate(request(), [
             'name' => ['required', 'max:100'],
             'slug' => ['unique:pages', 'required', 'max:50']
@@ -46,12 +51,18 @@ class PageController extends Controller
 
     public function edit($id)
     {
+        // Only Admins
+        if(!$this->hasrole('Admin')) { return redirect('/'); }
+
         $page = Page::find($id);
         $categories = Category::all();
         return view('pages/edit', compact('page', 'categories'));
     }
     public function update(Request $request, $id)
     {
+        // Only Admins
+        if(!$this->hasrole('Admin')) { return redirect('/'); }
+
         $page = Page::find($id);
 
         $this->validate(request(), [
@@ -64,6 +75,9 @@ class PageController extends Controller
     }
     public function destroy($id)
     {
+        // Only Admins
+        if(!$this->hasrole('Admin')) { return redirect('/'); }
+        
         $page = Page::find($id);
         Page::destroy($page->id);
         return redirect()->action('PageController@index');
